! Move all button_strip arrays to the source
IMO, there's no point having the button_strip arrays (such as
$mod_buttons or $normal_buttons) in the templates. It makes more sense
to have them in their corresponding source files, as with the main menu
in 2.0. Themes rarely need to modify them and if they do, they can still
do so (changing the image instance (which doesn't actually do much, does
it?) and so on) by just using the "new" array.

Also made sure there are hooks for all of them. One may argue that some
of them probably will never be used, but I feel that the hook system
won't have reached its full potential until as much as possible is
editable via hooks, so why not play it safe?