A massive commit containing many theme enhancements, bugfixes and a couple of additions.
!enhancement - News fader now only "fades" if there are multiple items.
({js_dir}fader.js)
!enhancement - Search button vertically aligned.
({theme_dir}index.template.php, {css_dir}index.css)
!enhancement - Mark all messages as read alignment fix.
({css_dir}index.css)
!bugfix - When opt.bDisabled is false for the JumpTo JavaScript object,
there is a "0" right before the select name. <select 0name="whatever".
This is caused by: <select ' + (this.opt.bDisabled == true ?
'disabled="disabled" ' : 0) the else should be an empty string, rather
than an integer so nothing outputs to the markup.
!enhancement - CSS "rfix" class. This class should be utilized whenever
you are 'integrating' a header (cat_bar / title_bar) with a roundframe.
This class gets rid of the one-pixel space between the header and the
body that looks like crap. ({css_dir}index.css)
!enhancement - Admin welcome header margin resolved by removing margin
in CSS (admin.css), and adding "rfix" class to the roundframe element
in the actual template. ({theme_dir}admin.template.php,
{css_dir}admin.css)
!enhancement - Admin quick search CSS vertical-alignment fixes.
!enhancement - Posting Icons vertically aligned properly.
!enhancement - ?action=who has an inconsistency. The second page index
and "show" select box are outside the main content box, yet the top one
is inside of it. Moved the second one back into the main box and fixed
all alignment issues with the second box.
!enhancement - ?action=who the upper show box had some inconstancies
with alignment in comparison to the pages block.
!enhancement - Alignment fix with the unapprove image sprite on posts
"more" drop down.
!bugfix - The CSS "divider" class for dividing boards on the BoardIndex
was set to: .table_list .divider so it had no affect, it was designed
to add a divider between boards. It should be applied to: .table_list
.divider tr td.
!enhancement - Fixed alignment of unread post icons and mark all
messages as read. Why this issue was fixed twice is because the "first"
fix did not incorporate room for the divider which was broken.
!enhancement - Fixed board icon vertical align.
!bugfix - BoardIndex layout was off with headers and their board
content. This was a big BoardIndex layout.
!enhancement - Stats block of boards borders are now #ddd; as opposed
to #ccc which is not the colour of the other borders.
!bugfix - Undefined Index "rss_limit" in index.template.php. Simply
added && !empty($modSettings['rss_limit']) for the condition of the RSS
link being shown, since the variable is used if it's true, it needs
substance.
!enhancement - Added "short_description" for boards, and utilized
"short_description" for the title="" of child boards that are redirects
so that when you mouseover them, you also see a "short" description of
the redirect without having to enter the parent. I set the length on
shorten_subject(); to 128 for now, but can definitely be lowered later
if we feel necessary.
!bugfix - Added "Members Viewing Topic" block to topics, since it was
left out of topics.
(http://dev.simplemachines.org/mantis/view.php?id=5004)
!bugfix - Code blocks have a minimum height of 10em, which is too
large. By removing the height attribute in the CSS it allows the block
to shrink or expand based on how much is inputted.
(http://dev.simplemachines.org/mantis/view.php?id=5003)

{theme_dir} = ./Themes/default/
{css_dir} = ./Themes/default/css/
{js_dir} = ./Themes/default/scripts