Enhance registration integration hooks
The existing hook integrate_register is described as being able to do further validation [http://wiki.simplemachines.org/smf/Integration_hooks#integrate_register]

However, it can not return errors to the user in the same way the registerMember function does ~20 lines above.

This patch is a modified version of the patch run locally that allows an integration hook to return an array of error messages to return to the user.

Secondly, this patch adds a post registration hook, that can be used to store the SMF member ID in a seperate system before the actual account is activated [i.e. integrate_activate is called]

This for example allows for the registration template to be customised to ask the user for custom information, and store that information for later review during the admin approval/activation process.

Signed-off-by: Paul Richards <paul@minimoo.org>