Minor Syntax Fixes + CSS Bug
Quite a few variables that were being concatenated were not consistant.
Used periods when commas were definitely appropriate, and then back to
commas for the next variable, and so forth.

The textarea when modifying a boards' description needed a pixel
margin-top to fix a display issue that was cutting off the top of the
field.

Sign-Off: Matthew Kerle <creativewebdsign@gmail.com>