Undefined Error: current_pm
Prior to the commit, the conversation mode buttons were being
constructed even if there were no message (reply | delete). Simply
moving this condition block into the above block if (!empty($pms)) is
the most efficient fix.

Huge thanks to Akyhne for pointing this issue out (#236), and to The
Craw for beginning to work on resolving it.

Signed-Off: Matthew Kerle <creativewebdsign@gmail.com>