Redirect Childboards + Short Description
This is another piece of my "A massive commit" commit that adds short
descriptions of child boards that are also a redirect:

!enhancement - Added "short_description" for boards, and utilized
"short_description" for the title="" of child boards that are redirects
so that when you mouseover them, you also see a "short" description of
the redirect without having to enter the parent. I set the length on
shorten_subject(); to 128 for now, but can definitely be lowered later
if we feel necessary.