Who Action Tweaks
Another piece of the "A massive commit" commit.

!enhancement - ?action=who has an inconsistency. The second page index
and "show" select box are outside the main content box, yet the top one
is inside of it. Moved the second one back into the main box and fixed
all alignment issues with the second box.
!enhancement - ?action=who the upper show box had some inconstancies
with alignment in comparison to the pages block.

Signed-off-by: emanuele <emanuele45@gmail.com>