fix for 'Unable to verify referring url'?
Potential fix for this long-known issue.  A bit crude perhaps, but it
does seem to work.

Issue seems to be caused by cookie expiry and the subsequent user of
HTTP_REFERER .  The user-agent seems to randomly add 'www.' to URLs  -
it's browser and version dependant.

PHP manual:
'  'HTTP_REFERER'
The address of the page (if any) which referred the user agent to the
current page. This is set by the user agent. Not all user agents will
set this, and some provide the ability to modify HTTP_REFERER as a
feature. In short, it cannot really be trusted.'