Update Security.php
For a permanent ban the user is logged out and associated data is removed. This should also include destroying the session and unset openid which the Logout() function takes care of. One can likely also omit the require Subs-Auth.php command because the Logout() function also loads that file but I will leave it up to you guys to remove that if you want. 

Signed-off-by: -Underdog- github.underdog@gmail.com