SCE set.command deprecated for SCE v1.4
The set.command function is deprecated as of SCE v1.4 and has been removed from the SCE library of functions. 
Since this function no longer exists in the SCE infrastructure the javascript for adding the BBC toolbar completely crashes when the code attempts to use setCommand to add the "before" key from the array. This caused backward compatibility with mods that add BBC's via hooks to fail (not to mention the entire toolbar went null).

The newer SCE function that takes its place is named command.set which has a different syntax.
The edit shown here replaces the deprecated function with the new one and its proper syntax.

The report was filed here: https://github.com/SimpleMachines/SMF2.1/issues/1845

Signed-off-by: -Underdog- github.underdog@gmail.com