current edit only works with display attribute
Atm I am not able to get this to work as documented on the sceditor website.
Perhaps there is something I am overlooking and will figure out.
For now surrounding the bbcode in a span with a display:none element whilst surrounding the html with a div utilizing the css class that hides it in source mode seems to work for me. The html must use a div container to allow all other content (whereas span will not but is ok for the bbcode text). 

display:inline element for the div container should allow the possible html containers within to behave as intended while not creating one additional line break that normally shows by default behavior. 

Signed-off-by: -Underdog- github.underdog@gmail.com