Default avatar not showing in the topheader miniprofile #2360
This is to fix issue #2360 where default avatar not showing in the
topheader miniprofile.

https://github.com/SimpleMachines/SMF2.1/issues/2360

Fix is not at the theme level as suggested earlier but Subs.php. IMO
this will ensure that no bug for having no avatar since default avatar
is supposed to be always in place.

Do check.