QuickReply Code Moved To Function Template
This is to ease modification which moves the quickreply box. With this,
the quickreply code is now moved into a function template_quickreply().

This function thus allows quickreply box to be moved anywhere without
moving the whole quick reply code. They just to find and move
template_quickreply(); to where they want it to be.

Signed-off-by: ahrasis <ahrasis@gmail.com>