Info Center Permission
I think this may be a very useful addition to SMF 2.1. I believe it was
started by deprecated and is very useful to many. But the mod never come
to the permission level until Labradoodle created a permission version
of it.

In this commit, I also attempted to ensure that the permission is on by
default and as such three sql files are edited as well.

Signed-off-by: ahrasis <ahrasis@gmail.com>