Restoring Span For $button['title']
This PR is only a suggestion for easy customization from this topic:
http://www.simplemachines.org/community/index.php?topic=536271.0. The
text menu can be easily removed by css display:none with the link and
icons left intact if the span is restored. There is no other easy way to
customize removing the text from menu so far without the span. While I
am pulling this request, I will leave this totally to the SMF Team
discretion to decide.

Signed-off-by: "ahrasis" Ahmad Rasyid Ismail <ahrasis@gmail.com>