Add missing default entries to user_alerts_prefs
based on the Notifications Settings page.

This should fix some issues (Like there being no index of a member when someone is added as an buddy).
The following preferences were not on that page, and are not included (As they do not seem to use the default value of user '0' anyway.):
msg_auto_notify, alert_timeout, msg_notify_pref, msg_notify_type, msg_receive_body, pm_notify

Signed-off-by: Stefan <git@fearthe1337.com>