Restoring No Span For Text Menu
Someone recently suggested that merely using css is actually possible to
remove text from menu in smaller screen in the latest post in
http://www.simplemachines.org/community/index.php?topic=536271.msg3815386#msg3815386.
I have tested his suggestion and it surprisingly works flawlessly. Since
this span is a bloat if there is no need for it to remove the text menu,
I would recommend the older code for it be restored i.e. this span be
removed again. I also modified the suggested css code in that post to be
similar to current css and remove the my previous css in this PR as
well.

Signed-off-by: "ahrasis" Ahmad Rasyid Ismail <ahrasis@gmail.com>