Fixes broken 'Remove empty BBC' regex The previous regex didn't make sense. The problematic part was:
(?>(?R)|.)*?

This makes an atomic group that first tries to recursively match nothing (which of course fails), and then tries to match any character except a line break, which only fails if the BBCode actually IS completely empty. Then the group is lazily repeated. As a result, this line was actually removing all BBCodes that were NOT empty!

This new version simply looks for zero or more whitespace characters between the BBCode tags. This does what the previous version seems to have been intended to do.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>