Faster, more robust handling of BBCode parameters Fixes #3027 (alt parameter of img bbcode tag not always parsed properly)
Fixes #2973 more effectively than #2975 by removing the need for permute() and pc_next_permutation()

The way this works is fairly straightforward. Instead of using permutations to figure out what arbitrary order a set of parameters happens to be in, we just sort them into a stable order and then work with that. As a result, we can avoid all the permutation stuff (and therefore the danger of factorially increasing computation times). Moreover, the parameters are **always** identified correctly no matter what order the user entered them in.

Gory details:
- Just as we are finishing building the array of BBCodes, ksort each BBCode's parameters sub-array (if present) in order to make sure they are arranged alphabetically by key name. Doing this job at this stage makes sure we only have to do it once.
- When it is time to parse the parameters the user entered, we do the following:
	1. Extract the string of parameters from the opening BBCode tag
	2. Loop through the expected parameter values for this tag and build two arrays. One is the venerable regular expression we've long used to check for parameter conformance. The other is a list of keys (plus equals signs) that we'll use to preg_split the parameter string we got in step 1.
	3. Use preg_split to split the param string into an array of params. Then sort the array of params alphabetically and implode it back into a string.
	4. Use the regex we made in step 2 to see if the parameter string matches our expectations for this BBCode. Because both the expected params and the given params have been sorted alphabetically, we only need to check once: it either matches, or it doesn't.
	5. Assuming the regex matched, proceed to handle the matches the same way we always have.

TODO: Fortify the string extraction code so that it can handle ] characters inside the parameter values.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>