Allow [code] inside block level BBC (e.g. lists and tables)
Previously the preparser split the message at every `[code]` and `[/code]`, parsed BBC only within the parts outside the `[code]...[/code]`, and reassembled the string at the end. Now we substitute the content of a `[code]...[/code]` with a placeholder, and then restore the original content at the end. This allows the rest of the BBC preparsing to operate on the entire $message at once, and therefore allows `[code]...[/code]` to remain in place during preparsing and thus to be embedded within list items, table cells, etc.

Fixes #3106

Signed-off-by: Jon Stovell <jonstovell@gmail.com>