Prevents resumed downloads from incrementing the download counter
If the client requests the download in multiple parts, we don't want each request to count as another complete download. With this change, we only update when the requested range includes the start of the file or doesn't specify a range at all.

Also makes sure we don't increment the counter unless we actually serve up some data. Previously, we were incrementing it even if we only emitted a 304 or 404 header.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>