Implements boardurl token substitution
As discussed in #3726, this commit implements seamless substitution of the value of $boardurl with a token when reading and writing to the database.

This approach solves a few problems:

* Activating or deactivating SSL will not break existing URLs in settings
* Changing the domain or path of the forum will not break existing URLs in settings
* Quotes and links in posts that refer to other posts will always use the proper value of $boardurl

It works by (1) replacing all occurrences of the value of $boardurl with the literal string `{$boardurl}` just before writing in smf_db_query(), and then (2) performing the opposite operation immediately after reading using smf_db_fetch_assoc() or smf_db_fetch_row().

Note:
1. During a fresh install, we no longer replace `{$boardurl}` in the install SQL files with the value of $boardurl, because we want to use that token in the database now.
2. During an upgrade to 2.1, we replace all literal references to the value of $boardurl in posts and personal messages with the token.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>