More rss improvements (#4030)
* Adds language declaration to RSS and Atom feeds

Signed-off-by: Jon Stovell <jonstovell@gmail.com>

* Uses value of $settings['og_image'] for feed icon image, if available

Signed-off-by: Jon Stovell <jonstovell@gmail.com>

* Appends board parameter to generated RSS links when appropriate

Signed-off-by: Jon Stovell <jonstovell@gmail.com>

* Sets default feed format to RSS2 instead of SMF's proprietary format
After all, RSS2 is used a lot, whereas our proprietary format is used by no one except... I dunno, maybe some crazy unknown mod floating out there somewhere in the ether.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>

* Tidies some repetitive code

Signed-off-by: Jon Stovell <jonstovell@gmail.com>

* Uses a less confusing variable name

Signed-off-by: Jon Stovell <jonstovell@gmail.com>

* Documentation corrections

Signed-off-by: Jon Stovell <jonstovell@gmail.com>

* Uses the tag URI scheme to generate GUIDs for feed items
Including the topic id in the GUID for a recent post meant that if the topic of the message changed, e.g. by splitting or merging topics, the GUID would change. That's bad.
Using the tag URI scheme lets us avoid that problem. This isn't quite compliant with the best practices, in that we make up the GUID on the fly every time, which means that if the forum changes its domain name, the GUID will change. But the domain name is a whole lot more stable than the other stuff we were using, so this is a significant improvement.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>

* Fixes a copy/paste error from a previous commit

Signed-off-by: Jon Stovell <jonstovell@gmail.com>

* Makes links for feed autodiscovery conform to standards

Signed-off-by: Jon Stovell <jonstovell@gmail.com>

* Fixes typo with adding board to feed source url

Signed-off-by: Jon Stovell <jonstovell@gmail.com>