Adds basic support for a float BBCode
Two forms are possible (with `left` or `right` available in either form):

1. `[float=left]example content goes here[/float]`
2. `[float=left max=33%]example content goes here[/float]`

The first form is straightforward. It floats the content to the left or the right based on the value entered after the = sign. The second form adds the ability to specify a max-width for the floating div. In this case, the max-width for the floating will be 33%. Note that on narrow screens, this setting will be overridden by responsive.css

Signed-off-by: Jon Stovell <jonstovell@gmail.com>