Correctly handles setting the displayed name during registration
0f46853 introduced a bug with allowing guests to set their displayed name during registration. Specifically, replacing `!empty($modSettings['allow_editDisplayName'])` with `allowedTo('profile_displayed_name')` meant that guests in the process of registering could never meet the condition, since `allowedTo('profile_displayed_name')` will always return false for guests.

This code fixes the problem by checking whether regular members have the 'profile_displayed_name_own' permission. If so, then the guest will be allowed to change their displayed name as soon as they complete registration anyway, so we allow it to be set during registration.

Of course, all of this only comes up for themes that offer a field to set the displayed name on the registration form (or if the user manually appends `real_name=foo` to the URL parameters). So to test this bug and its solution, you'll want to add an input field to the registration template for `$_POST['real_name']`.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>