Clean up some Serialized strings
For TFA which was introduced in 2.1, we shouldn’t need to keep the backup serialized fallback for the cookie.  2.1 is in dev and the cookies should all be expired.
A slight naming change for Logging.php to not say serialized
A possibly unused section of code we are making sure to convert it over to json for a charset issue
Remove a usage of serialized in Subs for timezone data

Signed-off-by: jdarwood007 <unmonitored+github@sleepycode.com>