Remembers member's real group in loadMemberData and loadMemberContext
Because $user_profile[$id]['member_group'] gets overwritten for moderators in loadMemberData(), there has been no easy way, short of a new query, to get the name of a moderator's primary membergroup. (The id number of the moderator's true primary group is retained, but not the group's name.)

With this change, $user_profile[$id]['primary_group'] and $memberContext[$user]['primary_group'] keep track of the name of member's true primary group.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>