Fixes/edits for Subs-Post.php
-  create empty arrays prior to attempting to use them (using a one
liner)
-  parse_bbc(false) always returns an empty string therefore do not
attempt to use it as an array in a for loop

If this (the 2nd edit) is supposed to be an attempt to filter
$code['tag'] through the parse_bbc function then it should be done
within the loop itself. If this is the intention then either change the
code to do that or put a notation in this PR and I will make it as such.

Signed-off-by: Underdog-01 <github.underdog@gmail.com>