Adds any/all logic to allowedTo() and isAllowedTo()
Normally, `allowedTo('post_new', $boards)` will return true if the current user can make new posts in all of the boards listed in $boards. If the $any parameter is set to true, as in `allowedTo('post_new', $boards, true)`, the function will return true if the current user can make new posts in at least one of the boards listed in $boards.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>