Uses filename and content to create unique hash when minimizing files (#4639)
By using the filename and a hash of the file's content, we can ensure that if the content changes in a file to be minimized, we create a new minimized file with a new name to reflect that.
The reason to do this is to prevent browsers from using obsolete cached versions of our minified files. Without this change, if a mod altered the content of an existing CSS or JS file, our minimized file's name would include the same hash that it did before the change. The browser could then conclude that it should use its out of date, locally cached version of the minimized file rather than fetching the current version. With this change, that will no longer happen. If the content of one of the original CSS or JS file has been changed, the filename of our minimized file will change to reflect that, prompting the browser to fetch the current version if necessary.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>