Upgrade can migrate our Settings file to a new version
The upgrader allows the migration our Settings.php file to a new version, if determined by a preset conditions that it is allowed to.  These conditions looks for initially integrations and advanced configuration setups and if found we do not allow the migration.  It then determines if variables are missing in the Settings.php that are added in new versions and if so suggests a migration.

The option is configurable during upgrade and can be overridden by the admin.  A copy of the Settings is made called Settings_org.php prior to the migration.  However is upgrade.php is ran multiple times and the option checked multiple times, it is possible to loose the original Settings prior to any migration.

This may require additional information and Settings.php example data to reliably determine if a migration is safe.

Signed-off-by: jdarwood007 <unmonitored+github@sleepycode.com>