Decouple the text statistics from the template
This also modifies
Create an array of the stats from the source side. This has an added benefit of mod authors being able to hook into this without modifying the template directly, via the integrate_profile_stats hook which got a new argument. Another benefit is that I've added url support so the stats can now link to other pages.

Signed-off-by: NanoSector <rick.2889@gmail.com>