change to select profile field broke backward compatibility
Recent changes to the profile fields template broke some backward compatibility and restricted select entries to non-numeric.

Imo non-numeric only for select entries was not necessary.
I suggest having devs use a disabled array to add/restrict disabled options.
Also having the name blank is not proper HTML.. it is supposed to be disabled with a proper name entry.

Signed-off-by: Underdog-01 <github.underdog@gmail.com>