Fixes awkward handling of legacy BBC
With these changes, legacy BBC can never be used in new posts, but they will continue to work if they are already present in pre-existing posts. This is because preparsecode() will strip them out (even if they are enabled in the admin settings), whereas parse_bbc() will treat them normally.

To make the difference between normal BBC and legacy BBC clearer to the admin, the two groups have been visually separated on the admin's BBC settings page and some help text has been added. This is purely a UI matter. There are no changes to how information about disabled BBC is stored in the database.

Also, the legacy BBC are now disabled by default for new installs, but are left alone during upgrades.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>