Always parses BBC using the recipent's language in email notifications Doing this requires using `parse_bbc()` in the langauge of the recipient of the email notification rather than in the language of the current user.
Main changes:
- `parse_bbc()` can now remember versions of the $bbc_codes array for multiple languages.
- For new post notifications, the body text of the message is now censored and parsed in CreatePost-Notify.php rather than in Subs-Post.php's `sendNotifications()`. This allows us to tailor the output of `parse_bbc()` depending on the recipient's language.
- Similarly, some code has been moved around and modified in the `sendpm()` function of Subs-Post.php.

TODO: This code temporarily resets `$txt` to a blank slate, populates it with the contents of the index language file for the recipient's language, and the restores `$txt` to its original state when we are done. This is fine for parsing standard BBC, but it will know nothing about `$txt` strings stored in other files, e.g. in those added by mods. Should probably do something about that...

Fixes #1407 (which dates all the way back to 2006: http://dev.simplemachines.org/mantis/view.php?id=52)

Signed-off-by: Jon Stovell <jonstovell@gmail.com>