Improves functionality of integrate_createAttachment hook
Previously, the integrate_createAttachment hook could be used to change existing values in $attachmentOptions, but it couldn't be used to handle custom values and columns. As a result, mods would be forced to make file changes if they wanted to populate a custom column they'd added to the attachments table. Now they can do that via the hook.

For safety's sake, these changes also add measures to prevent mods from breaking the standard behaviour if they screw up.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>