Add some hooks to our create_control_verification function, so mods may add additional verification options easily.
3 Hooks are added:
integrate_create_control_verification_pre - Intended for loading up additional settings or preprocessing.  Such as calling any additional classes/functions needed, getting things ready.
integrate_create_control_verification_test - When we are doing the test to see if they pass verification.  Only called during the post/submitting of a form.  We can return any errors here in the form of a text string name.
integrate_create_control_verification_refresh - If a request was made to refresh the verification options.  Can be used to let the mods know to reset options (or not).
integrate_create_control_verification_post - When we are done with the process.  Can be used to do cleanup and final checks before exiting out.