Alerts Configuration generates errors
When loading the Alerts configuration in ACP errors can be generated because we attempt to use $cur_profile, which from previous loadThemeOptions had specified 0 for the $memID, which results in a invalid $cur_profile.  To work around this we simply are adding a way to bypass the usage of $cur_profile and load Theme options from id_theme = 0.

This may need more testing than I gave it.