Better detect a working temp directory
Systems using open_basedir restrictions may receive errors with sys_get_temp_dir as it may not be properly configured to use a safe workable directory.  As well its possible that the temp directory may be readonly.  We also check if the directory is writable.

As a fall back we resort to sys_get_temp_dir again for lack of a 100% safe directory.