Undesirable results when a file doesn't exist when loading css or js.
If you call loadCSS('foo.bar'), the HTML output will be:
	<link rel="stylesheet" href="?smf21rc2_#TIMESTAMP#">

This is wrong.  This is produced because we only check that $fileName is not empty, not $fileUrl.  I'm not 100% sure if the intention to be able to call these without a valid file is intended, such as in the case of wanting to trigger the updateSettings call.  So instead of return, I just modified the check.