Fix issues with auth_secret missing in Settings.php
First we attempt to fix this during upgrades, but also allow the admin to randomize the auth_secret during upgrade.

Second attempt is during normal usage.  We log messages indicating it is missing but only once every 15 minutes.  For admins we issue a security warning as well.

As a fallback, we set secret_auth to a static value.  This weakens security, but at least allows logins to occur normally.

This also fixes some editor handling issues in Subs-Admin.php with regards to ?>
As well we have updated updateDbLastError to return a bool on success/failure and optionally updates the $db_last_error into the current context, allowing if needed for other parts of the system or 3rd party APIs to recognize we have an error that was written in the current context.